printformatter module testing
