+++ lib/print_formatter.rb
@@ -1,0 +1,30 @@
+
+
+module PrintFormatter
+
+  def headers
+    ["LAST NAME", "FIRST NAME", "EMAIL", "ZIPCODE", "CITY", "STATE", "ADDRESS", "PHONE", "DISTRICT"]
+  end
+
+  def create_attendee_array
+    queue.map do |attendee|
+      [attendee.last_name, attendee.first_name, attendee.email_address, attendee.zipcode, attendee.city, attendee.state, attendee.street, attendee.homephone, "DISTRICT"]
+    end
+  end
+
+  def create_line(words, length)
+    words.map do |word| word.ljust(length)
+    end.join(" ")
+  end
+
+  def find_max_characters
+    create_attendee_array.flatten.reduce(0) do |length, string|
+      if string.length > length
+        string.length
+      else
+        length
+      end
+    end
+  end
+
+end
